--Promote kinematic properties to base physics object
