+++ src/esp/physics/PhysicsObjectBase.h
@@ -8,6 +8,7 @@
 #include <Corrade/Containers/Optional.h>
 #include <Corrade/Containers/Reference.h>
 #include "esp/assets/ResourceManager.h"
+#include "esp/core/RigidState.h"
 
 /** @file
  * @brief Class @ref esp::physics::PhysicsObjectBase is the base class for any
@@ -113,7 +114,238 @@ class PhysicsObjectBase : public Magnum::SceneGraph::AbstractFeature3D {
    */
   const scene::SceneNode& getSceneNode() const { return node(); }
 
+  /**
+   * @brief Check whether object is being actively simulated, or sleeping.
+   * Kinematic objects are always active, but derived dynamics implementations
+   * may not be.  NOTE: no active objects without a physics engine...
+   * (kinematics don't count)
+   * @return true if active, false otherwise.
+   */
+  virtual bool isActive() const { return false; }
+
+  /**
+   * @brief Set an object as being actively simulated rather than sleeping.
+   * Kinematic objects are always active, but derived dynamics implementations
+   * may not be.
+   */
+  virtual void setActive() {}
+
+  /**
+   * @brief Set the light setup of this rigid.
+   * @param lightSetupKey @ref gfx::LightSetup key
+   */
+  void setLightSetup(const std::string& lightSetupKey) {
+    gfx::setLightSetupForSubTree(node(), lightSetupKey);
+  }
+
+  // ==== Transformations ===
+
+  /** @brief Set the 4x4 transformation matrix of the object kinematically.
+   * Calling this during simulation of a @ref MotionType::DYNAMIC object is not
+   * recommended.
+   * @param transformation The desired 4x4 transform of the object.
+   */
+  virtual void setTransformation(const Magnum::Matrix4& transformation) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().setTransformation(transformation);
+      syncPose();
+    }
+  }
+
+  virtual Magnum::Matrix4 getTransformation() const {
+    return node().transformation();
+  }
+
+  /** @brief Set the 3D position of the object kinematically.
+   * Calling this during simulation of a @ref MotionType::DYNAMIC object is not
+   * recommended.
+   * @param vector The desired 3D position of the object.
+   */
+  virtual void setTranslation(const Magnum::Vector3& vector) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().setTranslation(vector);
+      syncPose();
+    }
+  }
+
+  virtual Magnum::Vector3 getTranslation() const {
+    return node().translation();
+  }
+
+  /** @brief Set the orientation of the object kinematically.
+   * Calling this during simulation of a @ref MotionType::DYNAMIC object is not
+   * recommended.
+   * @param quaternion The desired orientation of the object.
+   */
+  virtual void setRotation(const Magnum::Quaternion& quaternion) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().setRotation(quaternion);
+      syncPose();
+    }
+  }
+
+  virtual Magnum::Quaternion getRotation() const { return node().rotation(); }
+
+  /**
+   * @brief Get the rotation and translation of the object.
+   */
+  virtual core::RigidState getRigidState() {
+    return core::RigidState(node().rotation(), node().translation());
+  };
+
+  /**
+   * @brief Set the rotation and translation of the object.
+   */
+  virtual void setRigidState(const core::RigidState& rigidState) {
+    setTranslation(rigidState.translation);
+    setRotation(rigidState.rotation);
+  };
+
+  /** @brief Reset the transformation of the object.
+   * !!NOT IMPLEMENTED!!
+   */
+  virtual void resetTransformation() {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().resetTransformation();
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the 3D position of the object kinematically by translation.
+   * Calling this during simulation of a @ref MotionType::DYNAMIC object is not
+   * recommended.
+   * @param vector The desired 3D vector by which to translate the object.
+   */
+  virtual void translate(const Magnum::Vector3& vector) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().translate(vector);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the 3D position of the object kinematically by translation
+   * with a vector defined in the object's local coordinate system. Calling this
+   * during simulation of a @ref MotionType::DYNAMIC object is not recommended.
+   * @param vector The desired 3D vector in the object's ocal coordiante system
+   * by which to translate the object.
+   */
+  virtual void translateLocal(const Magnum::Vector3& vector) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().translateLocal(vector);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying an
+   * axis-angle rotation to it. Calling this during simulation of a @ref
+   * MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   * @param normalizedAxis The desired unit vector axis of rotation.
+   */
+  virtual void rotate(const Magnum::Rad angleInRad,
+                      const Magnum::Vector3& normalizedAxis) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotate(angleInRad, normalizedAxis);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying an
+   * axis-angle rotation to it in the local coordinate system. Calling this
+   * during simulation of a @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   * @param normalizedAxis The desired unit vector axis of rotation in the local
+   * coordinate system.
+   */
+  virtual void rotateLocal(const Magnum::Rad angleInRad,
+                           const Magnum::Vector3& normalizedAxis) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateLocal(angleInRad, normalizedAxis);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying a
+   * rotation to it about the global X axis. Calling this during simulation of a
+   * @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   */
+  virtual void rotateX(const Magnum::Rad angleInRad) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateX(angleInRad);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying a
+   * rotation to it about the global Y axis. Calling this during simulation of a
+   * @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   */
+  virtual void rotateY(const Magnum::Rad angleInRad) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateY(angleInRad);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying a
+   * rotation to it about the global Z axis. Calling this during simulation of a
+   * @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   */
+  virtual void rotateZ(const Magnum::Rad angleInRad) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateZ(angleInRad);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying a
+   * rotation to it about the local X axis. Calling this during simulation of a
+   * @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   */
+  virtual void rotateXLocal(const Magnum::Rad angleInRad) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateXLocal(angleInRad);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying a
+   * rotation to it about the local Y axis. Calling this during simulation of a
+   * @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   */
+  virtual void rotateYLocal(const Magnum::Rad angleInRad) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateYLocal(angleInRad);
+      syncPose();
+    }
+  }
+
+  /** @brief Modify the orientation of the object kinematically by applying a
+   * rotation to it about the local Z axis. Calling this during simulation of a
+   * @ref MotionType::DYNAMIC object is not recommended.
+   * @param angleInRad The angle of rotation in radians.
+   */
+  virtual void rotateZLocal(const Magnum::Rad angleInRad) {
+    if (objectMotionType_ != MotionType::STATIC) {
+      node().rotateZLocal(angleInRad);
+      syncPose();
+    }
+  }
+
+  /** @brief Store whatever object attributes you want here! */
+  esp::core::Configuration::ptr attributes_{};
+
  protected:
+  /** @brief Used to synchronize other simulator's notion of the object state
+   * after it was changed kinematically. Must be called automatically on
+   * kinematic updates.*/
+  virtual void syncPose() { return; }
+
   /** @brief An assignable name for this object.
    */
   std::string objectName_;
