+++ src/esp/physics/objectWrappers/ManagedPhysicsObjectBase.h
@@ -36,6 +36,11 @@ class AbstractManagedPhysicsObject : public esp::core::AbstractManagedObject {
   ~AbstractManagedPhysicsObject() override = default;
 
   /**
+   * @brief Test whether this wrapper's object still exists.
+   */
+  bool isAlive() { return !weakObjRef_.expired(); }
+
+  /**
    * @brief Get this managed object's class.  Should only be set from
    * constructor. Used as key in constructor function pointer maps in Managed
    * Container.
@@ -87,10 +92,166 @@ class AbstractManagedPhysicsObject : public esp::core::AbstractManagedObject {
     }
   }
 
-  /**
-   * @brief Test whether this wrapper's object still exists.
-   */
-  bool isAlive() { return !weakObjRef_.expired(); }
+  bool isActive() {
+    if (auto sp = this->getObjectReference()) {
+      return sp->isActive();
+    } else {
+      return false;
+    }
+  }  // isActive()
+
+  void setActive() {
+    if (auto sp = this->getObjectReference()) {
+      sp->setActive();
+    }
+  }  // activate()
+
+  void setLightSetup(const std::string& lightSetupKey) {
+    if (auto sp = this->getObjectReference()) {
+      sp->setLightSetup(lightSetupKey);
+    }
+  }  // setLightSetup
+
+  scene::SceneNode* getSceneNode() {
+    if (auto sp = this->getObjectReference()) {
+      return &const_cast<scene::SceneNode&>(sp->getSceneNode());
+    } else {
+      return nullptr;
+    }
+  }  // getSceneNode
+
+  esp::core::Configuration::ptr userAttributes() {
+    if (auto sp = this->getObjectReference()) {
+      return sp->attributes_;
+    } else {
+      return nullptr;
+    }
+  }
+
+  // ==== Transformations ===
+
+  Magnum::Matrix4 getTransformation() const {
+    if (auto sp = this->getObjectReference()) {
+      return sp->getTransformation();
+    } else {
+      return Magnum::Matrix4{};
+    }
+  }  // getTransformation
+
+  void setTransformation(const Magnum::Matrix4& transformation) {
+    if (auto sp = this->getObjectReference()) {
+      sp->setTransformation(transformation);
+    }
+  }  // setTransformation
+
+  Magnum::Vector3 getTranslation() const {
+    if (auto sp = this->getObjectReference()) {
+      return sp->getTranslation();
+    } else {
+      return Magnum::Vector3{};
+    }
+  }  // getTranslation
+
+  void setTranslation(const Magnum::Vector3& vector) {
+    if (auto sp = this->getObjectReference()) {
+      sp->setTranslation(vector);
+    }
+  }  // setTranslation
+
+  Magnum::Quaternion getRotation() const {
+    if (auto sp = this->getObjectReference()) {
+      return sp->getRotation();
+    } else {
+      return Magnum::Quaternion{};
+    }
+  }  // getTranslation
+  void setRotation(const Magnum::Quaternion& quaternion) {
+    if (auto sp = this->getObjectReference()) {
+      sp->setRotation(quaternion);
+    }
+  }  // setRotation
+
+  core::RigidState getRigidState() {
+    if (auto sp = this->getObjectReference()) {
+      return sp->getRigidState();
+    } else {
+      return core::RigidState{};
+    }
+  }  // getRigidState()
+
+  void setRigidState(const core::RigidState& rigidState) {
+    if (auto sp = this->getObjectReference()) {
+      sp->setRigidState(rigidState);
+    }
+  }  // setRigidState
+
+  void resetTransformation() {
+    if (auto sp = this->getObjectReference()) {
+      sp->resetTransformation();
+    }
+  }  // resetTransformation
+
+  void translate(const Magnum::Vector3& vector) {
+    if (auto sp = this->getObjectReference()) {
+      sp->translate(vector);
+    }
+  }  // translate
+
+  void translateLocal(const Magnum::Vector3& vector) {
+    if (auto sp = this->getObjectReference()) {
+      sp->translateLocal(vector);
+    }
+  }  // translateLocal
+
+  void rotate(const Magnum::Rad angleInRad,
+              const Magnum::Vector3& normalizedAxis) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotate(angleInRad, normalizedAxis);
+    }
+  }  // rotate
+
+  void rotateLocal(const Magnum::Rad angleInRad,
+                   const Magnum::Vector3& normalizedAxis) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateLocal(angleInRad, normalizedAxis);
+    }
+  }  // rotateLocal
+
+  void rotateX(const Magnum::Rad angleInRad) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateX(angleInRad);
+    }
+  }  // rotateX
+
+  void rotateY(const Magnum::Rad angleInRad) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateY(angleInRad);
+    }
+  }  // rotateY
+
+  void rotateZ(const Magnum::Rad angleInRad) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateZ(angleInRad);
+    }
+  }  // rotateZ
+
+  void rotateXLocal(const Magnum::Rad angleInRad) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateXLocal(angleInRad);
+    }
+  }  // rotateXLocal
+
+  void rotateYLocal(const Magnum::Rad angleInRad) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateYLocal(angleInRad);
+    }
+  }  // rotateYLocal
+
+  void rotateZLocal(const Magnum::Rad angleInRad) {
+    if (auto sp = this->getObjectReference()) {
+      sp->rotateZLocal(angleInRad);
+    }
+  }  // rotateZLocal
 
  protected:
   /**
